just a rearrangement
